refactor: move files in utils to config dir
