+++ packages/build-plugin-rax-component/src/config/getDemos.js
