+++ packages/build-plugin-rax-component/src/config/parseMarkdown.js
