+++ packages/build-plugin-rax-component/src/config/weex/getDev.js
@@ -2,7 +2,7 @@ const path = require('path');
 const { hmrClient } = require('rax-compile-config');
 const WeexFrameworkBanner = require('../../plugins/WeexFrameworkBannerPlugin');
 const getBaseWebpack = require('../getBaseWebpack');
-const getDemos = require('../utils/getDemos');
+const getDemos = require('../getDemos');
 
 module.exports = (context) => {
   const config = getBaseWebpack(context);
